requirements.txt
# Core
python-binance>=1.0.19
websockets>=12.0
aiohttp>=3.9.0

# Multi-Exchange
pybit>=5.6.0
python-okx>=0.2.0

# Web dashboard
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Performance
uvloop>=0.19.0; sys_platform != 'win32'
orjson>=3.9.0

# Security
PyJWT>=2.8.0
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4

# Data processing
pandas>=2.1.0
numpy>=1.26.0
polars>=0.20.0

# Technical analysis
ta>=0.11.0
pandas-ta>=0.3.14b

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis>=5.0.0
hiredis>=2.3.0

# Caching
cachetools>=5.3.0

# Logging
loguru>=0.7.0

# Monitoring
prometheus-client>=0.19.0

# Telegram
python-telegram-bot>=20.7

# ML/Optimization
scikit-learn>=1.3.0
optuna>=3.5.0
joblib>=1.3.0

# HTTP client
httpx>=0.26.0

# Jupyter (for analysis)
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Scheduling
apscheduler>=3.10.0

# Retry logic
tenacity>=8.2.0
